+++ CHANGES.md
@@ -3,6 +3,7 @@
   * Add base to filter\intval
   * Make filter\boolval php <5.5 compatible
   * New filter\sanitize
+  * Make Invalid compatible with \Exception
 
 ### 1.0.0-RC1 (2014-08-07)